Followed Coach's advices
