+++ src/C2L2/Hogwarts.java
@@ -1,22 +1,6 @@
 package C2L2;
 
-public class Hogwarts {
-
-    public static void compareHogwartians(Hogwarts h1, Hogwarts h2) {
-        int r1 = h1.rateAsHogwartian();
-        int r2 = h2.rateAsHogwartian();
-        System.out.print(h1.getFullName() + " with " + r1 + " points as Hogwartian is ");
-        if (r1 == r2) {
-            System.out.print("same as ");
-        }
-        if (r1 > r2) {
-            System.out.print("better than ");
-        }
-        if (r1 < r2) {
-            System.out.print("is worse than ");
-        }
-        System.out.println(h2.getFullName() + " with " + r2 + " points\n");
-    }
+public abstract class Hogwarts {
 
     private final static String schoolName = "Hogwarts School of Witchcraft and Wizardry";
     private String name;
